Support multiple repos

+++ diff_local.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-LOCKDIR=/home/mozzie/gitlock
-REPO=/home/mozzie/svm/
+LOCKDIR=/data/scummvm/gitmailer/gitlock
+REPO=/data/scummvm/gitmailer/repo
 GIT=/usr/bin/git
 
 while ! mkdir $LOCKDIR; do
@@ -12,11 +12,16 @@ if [ ! -n "$1" ]
 then
   exit 0
 fi
+if [ ! -n "$2" ]
+then
+  exit 0
+fi
 
-REV=$1
+REPONAME=$1
+REV=$2
 
-cd $REPO
-$GIT fetch -a > /dev/null
-$GIT diff -p --no-color  $REV~1..$REV --
+cd $REPO/$REPONAME
+$GIT fetch -a -q > /dev/null
+$GIT diff -p --no-color  $REV~1..$REV -- | head -n 10000
 
 rmdir $LOCKDIR
